feat(columbia_merger): Add possible permutations for cluster source
Move functions to columbia_utils.py
